+++ src/App.svelte
@@ -36,6 +36,7 @@
   import SourceSwitcher from './SourceSwitcher.svelte'
   import ProfileSelect from './ProfileSelect.svelte'
   import SceneCollectionSelect from './SceneCollectionSelect.svelte'
+  import Randomizer from './Randomizer.svelte';
 
   onMount(async () => {
     if ('serviceWorker' in navigator) {
@@ -93,15 +94,7 @@
     window.sendCommand = sendCommand
   })
 
-  const delay = ms => new Promise(res => setTimeout(res, ms));
-
   // State
-  let statusrandom = false;
-  let previousrandom;
-  let timeoutrandom;
-  let delaytime = 5;
-  let transitionsmode = '';
-  let transitions = []
   let connected
   let heartbeat = {}
   let heartbeatInterval
@@ -207,80 +200,6 @@
     await sendCommand('ResumeRecord')
   }
 
-  async function setMode() {
-    console.log(mode);
-  }
-
-  async function setdelay() {
-    //delaytime = 1;
-    console.log(delaytime);
-  }
-
-  async function RandomScene() {
-
-    if (!statusrandom) {
-
-      console.log('start --> randomscene');
-      statusrandom = true;
-      StartRandomScene()
-      //await sendCommand('BroadcastCustomMessage', { realm: 'randomscene', data: {randomscene: 'activate'}});
-
-    } else {
-
-      console.log('stop --> randomscene');
-      //await sendCommand('BroadcastCustomMessage', { realm: 'randomscene', data: {randomscene: 'deactivate'}});
-      statusrandom = false;
-    }
-
-  }
-
-  async function StartRandomScene() {
-
-    while (true) {
-
-      if (statusrandom) {
-        let randomness = Math.floor(Math.random() * 4)
-        let timeoutrandom = (randomness + parseInt(delaytime)) * 1000;
-        console.log('delay set',timeoutrandom/1000, 'randomness', randomness, 'delaytime', delaytime);
-        await delay(timeoutrandom);
-        console.log('delay finished');
-        GetRandomScene();
-
-      } else {
-
-        break;
-
-      }
-
-    }
-
-  }
-
-
-  async function GetRandomScene() {
-    
-    let data = await sendCommand('GetSceneList')
-
-    var item = data.scenes[Math.floor(Math.random()*data.scenes.length)];
-    if (!previousrandom) previousrandom = item;
-
-    while (item.sceneName == previousrandom.sceneName) {
-      var item = data.scenes[Math.floor(Math.random()*data.scenes.length)];
-    } 
-
-    // console.log(item);
-    //await sendCommand('SetCurrentScene', { 'scene-name': 'Cam' + previousrandom });
-
-    await sendCommand('SetCurrentPreviewScene', { 'sceneName': item.sceneName});
-    await sendCommand('SetCurrentSceneTransition', { transitionName: mode })
-    await sendCommand('TriggerStudioModeTransition')
-
-    previousrandom = item;
-
-    await sendCommand('BroadcastCustomEvent', { realm: 'randomscene', eventData: {randomscene: 'active'}});
-
-  }
-
   async function connect () {
     address = address || 'ws://localhost:4455'
     if (address.indexOf('://') === -1) {
@@ -296,14 +215,7 @@
       )
       console.log(
         `Connected to obs-websocket version ${obsWebSocketVersion} (using RPC ${negotiatedRpcVersion})`
-      )
-
-      let data;
-      data = await sendCommand('GetSceneTransitionList')
-      transitions = data.transitions || []
-
-      transitionsmode = data.transitions[0].transitionName;
-      
+      )      
     } catch (e) {
       console.log(e)
       errorMessage = e.message
@@ -487,28 +399,6 @@
               </button>
             {/if}
             
-            {#if statusrandom} 
-              <button class="button is-danger" on:click={RandomScene} title="Stop Random Scene" >
-              <span class="icon"><Icon path={mdiShuffle} /></span>
-              </button>
-            {:else}
-              <button class="button is-danger is-light" on:click={RandomScene} title="Random Scene" >
-              <span class="icon"><Icon path={mdiShuffle} /></span>
-              </button>
-            {/if}
-
-            <div class="button is-info is-light" style="margin: 0 .5rem .5rem 0;">
-              <input class="is-info" title="Change Delaytime" bind:value={delaytime} on:change={setdelay} style="line-height: 2.7em;width: 3em;text-align: center;border-top-style: hidden; border-right-style: hidden; border-left-style: hidden; border-bottom-style: hidden; background-color: #eff5fb;">
-            </div>
-
-            <div class="select" style="margin: 0 .5rem .5rem 0;">
-              <select bind:value={transitionsmode} title="Change Profile" on:change={setMode}>
-                {#each transitions as transition}
-                  <option value={transition.transitionName}>{transition.transitionName}</option>
-                {/each}
-              </select>
-            </div>
-
             {#if isVirtualCamActive}
               <button
                 class="button is-danger"
@@ -580,6 +470,8 @@
               </span>
               {#if replayError}<span>{replayError}</span>{/if}
             </button>
+
+            <Randomizer />
             <ProfileSelect />
             <SceneCollectionSelect />
             <button
